Move hybrid scan reader tests to a separate executable
