+++ cpp/tests/CMakeLists.txt
@@ -319,13 +319,21 @@ ConfigureTest(
   io/parquet_bloom_filter_test.cu
   io/parquet_chunked_reader_test.cu
   io/parquet_chunked_writer_test.cpp
-  io/parquet_experimental_reader_test.cpp
   io/parquet_common.cpp
   io/parquet_misc_test.cpp
   io/parquet_reader_test.cpp
   io/parquet_test.cpp
   io/parquet_v2_test.cpp
   io/parquet_writer_test.cpp
+  GPUS 1
+  PERCENT 30
+)
+ConfigureTest(
+  HYBRID_SCAN_TEST
+  io/experimental/hybrid_scan_filters_test.cpp
+  io/experimental/hybrid_scan_common.cpp
+  io/parquet_test.cpp
+  io/parquet_common.cpp
   GPUS 1
   PERCENT 30
 )
