+++ cpp/tests/io/experimental/hybrid_scan_common.cpp
@@ -1,0 +1,97 @@
+
+/*
+ * Copyright (c) 2025, NVIDIA CORPORATION.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "hybrid_scan_common.hpp"
+
+#include <cudf/io/parquet.hpp>
+#include <cudf/io/text/byte_range_info.hpp>
+#include <cudf/table/table.hpp>
+#include <cudf/types.hpp>
+#include <cudf/utilities/default_stream.hpp>
+#include <cudf/utilities/memory_resource.hpp>
+#include <cudf/utilities/span.hpp>
+
+cudf::host_span<uint8_t const> fetch_footer_bytes(cudf::host_span<uint8_t const> buffer)
+{
+  using namespace cudf::io::parquet;
+
+  constexpr auto header_len = sizeof(file_header_s);
+  constexpr auto ender_len  = sizeof(file_ender_s);
+  size_t const len          = buffer.size();
+
+  auto const header_buffer = cudf::host_span<uint8_t const>(buffer.data(), header_len);
+  auto const header        = reinterpret_cast<file_header_s const*>(header_buffer.data());
+  auto const ender_buffer =
+    cudf::host_span<uint8_t const>(buffer.data() + len - ender_len, ender_len);
+  auto const ender = reinterpret_cast<file_ender_s const*>(ender_buffer.data());
+  CUDF_EXPECTS(len > header_len + ender_len, "Incorrect data source");
+  constexpr uint32_t parquet_magic = (('P' << 0) | ('A' << 8) | ('R' << 16) | ('1' << 24));
+  CUDF_EXPECTS(header->magic == parquet_magic && ender->magic == parquet_magic,
+               "Corrupted header or footer");
+  CUDF_EXPECTS(ender->footer_len != 0 && ender->footer_len <= (len - header_len - ender_len),
+               "Incorrect footer length");
+
+  return cudf::host_span<uint8_t const>(buffer.data() + len - ender->footer_len - ender_len,
+                                        ender->footer_len);
+}
+
+cudf::host_span<uint8_t const> fetch_page_index_bytes(
+  cudf::host_span<uint8_t const> buffer, cudf::io::text::byte_range_info const page_index_bytes)
+{
+  return cudf::host_span<uint8_t const>(
+    reinterpret_cast<uint8_t const*>(buffer.data()) + page_index_bytes.offset(),
+    page_index_bytes.size());
+}
+
+std::vector<rmm::device_buffer> fetch_byte_ranges(
+  cudf::host_span<uint8_t const> host_buffer,
+  cudf::host_span<cudf::io::text::byte_range_info const> byte_ranges,
+  rmm::cuda_stream_view stream,
+  rmm::device_async_resource_ref mr)
+{
+  std::vector<rmm::device_buffer> buffers{};
+  buffers.reserve(byte_ranges.size());
+
+  std::transform(
+    byte_ranges.begin(),
+    byte_ranges.end(),
+    std::back_inserter(buffers),
+    [&](auto const& byte_range) {
+      auto const chunk_offset = host_buffer.data() + byte_range.offset();
+      auto const chunk_size   = byte_range.size();
+      auto buffer             = rmm::device_buffer(chunk_size, stream, mr);
+      CUDF_CUDA_TRY(cudaMemcpyAsync(
+        buffer.data(), chunk_offset, chunk_size, cudaMemcpyHostToDevice, stream.value()));
+      return buffer;
+    });
+
+  stream.synchronize_no_throw();
+  return buffers;
+}
+
+cudf::test::strings_column_wrapper constant_strings(cudf::size_type value)
+{
+  CUDF_EXPECTS(value >= 0 && value <= 9999, "String value must be between 0000 and 9999");
+
+  std::array<char, 5> buf;
+  auto elements =
+    thrust::make_transform_iterator(thrust::make_constant_iterator(value), [&buf](auto i) {
+      sprintf(buf.data(), "%04d", i);
+      return std::string(buf.data());
+    });
+  return cudf::test::strings_column_wrapper(elements, elements + num_ordered_rows);
+}
